fix: asset actions disclaimer position
