+++ src/components/organisms/AssetActions/Compute/index.tsx
@@ -59,13 +59,15 @@ export default function Compute({
   file,
   fileIsLoading,
   isConsumable,
-  consumableFeedback
+  consumableFeedback,
+  computeDisclaimerMessage
 }: {
   dtBalance: string
   file: FileMetadata
   fileIsLoading?: boolean
   isConsumable?: boolean
   consumableFeedback?: string
+  computeDisclaimerMessage?: string
 }): ReactElement {
   const { appConfig } = useSiteMetadata()
   const { accountId } = useWeb3()
@@ -482,7 +484,11 @@ export default function Compute({
           />
         </Formik>
       )}
-
+      {computeDisclaimerMessage && (
+        <div className={styles.disclaimer}>
+          <Alert state="info" text={computeDisclaimerMessage} />
+        </div>
+      )}
       <footer className={styles.feedback}>
         {isPublished && (
           <SuccessConfetti
