+++ src/components/organisms/AssetActions/Consume.module.css
@@ -16,3 +16,7 @@
   width: 100%;
   margin-top: calc(var(--spacer));
 }
+
+.disclaimer {
+  margin-top: var(--spacer);
+}
