+++ src/components/organisms/AssetActions/index.tsx
@@ -17,8 +17,38 @@ import axios from 'axios'
 import { getOceanConfig } from '../../../utils/ocean'
 import { useCancelToken } from '../../../hooks/useCancelToken'
 import { useIsMounted } from '../../../hooks/useIsMounted'
+import { graphql, useStaticQuery } from 'gatsby'
+
+const query = graphql`
+  query {
+    content: allFile(filter: { relativePath: { eq: "assetDisclaimer.json" } }) {
+      edges {
+        node {
+          childContentJson {
+            message
+          }
+        }
+      }
+    }
+  }
+`
+
+interface DisclaimerData {
+  content: {
+    edges: {
+      node: {
+        childContentJson: {
+          message: string
+        }
+      }
+    }[]
+  }
+}
 
 export default function AssetActions(): ReactElement {
+  const data: DisclaimerData = useStaticQuery(query)
+  const { message } = data.content.edges[0].node.childContentJson
+
   const { accountId, balance } = useWeb3()
   const { ocean, account } = useOcean()
   const { price, ddo, isAssetNetwork } = useAsset()
@@ -107,6 +137,7 @@ export default function AssetActions(): ReactElement {
       fileIsLoading={fileIsLoading}
       isConsumable={isConsumable}
       consumableFeedback={consumableFeedback}
+      computeDisclaimerMessage={message}
     />
   ) : (
     <Consume
@@ -117,6 +148,7 @@ export default function AssetActions(): ReactElement {
       fileIsLoading={fileIsLoading}
       isConsumable={isConsumable}
       consumableFeedback={consumableFeedback}
+      consumeDisclaimerMessage={message}
     />
   )
